refactor models.py into submodule
