+++ src/carapace/agent/loop.py
@@ -28,7 +28,7 @@ from pydantic_ai.messages import (
 from pydantic_ai.usage import UsageLimits
 
 from carapace.agent.tools import create_agent
-from carapace.models import Deps, TaskDone, TaskFailed
+from carapace.models.agent import Deps, TaskDone, TaskFailed
 from carapace.security.context import (
     AgentResponseEntry,
     ApprovalEntry,
