+++ src/carapace/channels/matrix/commands.py
@@ -2,7 +2,8 @@
 
 from __future__ import annotations
 
-from carapace.models import Deps, context_grants_session_summary
+from carapace.models.agent import Deps
+from carapace.models.skills import context_grants_session_summary
 from carapace.ws_models import CommandResult
