+++ src/carapace/jobs.py
@@ -9,7 +9,7 @@ from zoneinfo import ZoneInfo
 import yaml
 from croniter import croniter
 
-from carapace.models import JobCronTrigger, JobDefinition, JobsFile
+from carapace.models.jobs import JobCronTrigger, JobDefinition, JobsFile
 
 JobTriggerKind = Literal["api", "cron", "manual"]
