+++ src/carapace/models/__init__.py
@@ -1,0 +1,106 @@
+from __future__ import annotations
+
+from carapace.models.agent import Deps, TaskDone, TaskFailed
+from carapace.models.config import (
+    AgentConfig,
+    AvailableModelEntry,
+    CacheConfig,
+    CarapaceConfig,
+    ChannelsConfig,
+    Config,
+    CronChannelConfig,
+    CronJobConfig,
+    GitConfig,
+    MatrixChannelConfig,
+    MatrixTokenFile,
+    SandboxConfig,
+    Secret,
+    ServerConfig,
+    SessionCommitConfig,
+    SessionsConfig,
+    agent_available_model_entries,
+)
+from carapace.models.credentials import (
+    BitwardenCredentialBackendConfig,
+    CredentialBackendConfig,
+    CredentialMetadata,
+    CredentialRegistryProtocol,
+    CredentialsConfig,
+    FileCredentialBackendConfig,
+)
+from carapace.models.jobs import JobCronTrigger, JobDefinition, JobsFile
+from carapace.models.notifications import (
+    NotificationClientType,
+    NotificationFocusState,
+    NotificationPreferences,
+    NotificationsConfig,
+    NotificationSubscription,
+)
+from carapace.models.session import SessionAttributes, SessionBudget, SessionJobRunContext, SessionState
+from carapace.models.skills import (
+    ContextGrant,
+    SkillCarapaceConfig,
+    SkillCommandDecl,
+    SkillCredentialDecl,
+    SkillInfo,
+    SkillNetworkConfig,
+    context_grants_session_summary,
+)
+from carapace.models.tooling import (
+    ToolCallCallback,
+    ToolResult,
+    normalize_optional_tool_label,
+    normalize_tool_call_args,
+)
+
+__all__ = [
+    "AgentConfig",
+    "AvailableModelEntry",
+    "BitwardenCredentialBackendConfig",
+    "CacheConfig",
+    "CarapaceConfig",
+    "ChannelsConfig",
+    "Config",
+    "ContextGrant",
+    "CredentialBackendConfig",
+    "CredentialMetadata",
+    "CredentialRegistryProtocol",
+    "CredentialsConfig",
+    "CronChannelConfig",
+    "CronJobConfig",
+    "Deps",
+    "FileCredentialBackendConfig",
+    "GitConfig",
+    "JobCronTrigger",
+    "JobDefinition",
+    "JobsFile",
+    "MatrixChannelConfig",
+    "MatrixTokenFile",
+    "NotificationClientType",
+    "NotificationFocusState",
+    "NotificationPreferences",
+    "NotificationSubscription",
+    "NotificationsConfig",
+    "SandboxConfig",
+    "Secret",
+    "ServerConfig",
+    "SessionAttributes",
+    "SessionBudget",
+    "SessionCommitConfig",
+    "SessionJobRunContext",
+    "SessionState",
+    "SessionsConfig",
+    "SkillCarapaceConfig",
+    "SkillCommandDecl",
+    "SkillCredentialDecl",
+    "SkillInfo",
+    "SkillNetworkConfig",
+    "TaskDone",
+    "TaskFailed",
+    "ToolCallCallback",
+    "ToolResult",
+    "agent_available_model_entries",
+    "context_grants_session_summary",
+    "normalize_optional_tool_label",
+    "normalize_tool_call_args",
+]
