+++ src/carapace/session/archive.py
@@ -14,7 +14,8 @@ from loguru import logger
 from pydantic_ai.messages import ModelRequest, ModelResponse, TextPart, ThinkingPart, ToolCallPart, UserPromptPart
 
 from carapace.git.store import GitStore
-from carapace.models import SessionCommitConfig, SessionState
+from carapace.models.config import SessionCommitConfig
+from carapace.models.session import SessionState
 from carapace.session.manager import SessionManager
 
 ArchiveTrigger = Literal["manual", "autosave", "archive"]
