+++ src/carapace/session/commands.py
@@ -7,7 +7,8 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, Protocol
 
 from carapace.git.store import GitStore
-from carapace.models import Config, SkillInfo, context_grants_session_summary
+from carapace.models.config import Config
+from carapace.models.skills import SkillInfo, context_grants_session_summary
 from carapace.sandbox.manager import SandboxManager
 from carapace.session.manager import SessionManager
 from carapace.session.types import ActiveSession
