+++ src/carapace/session/engine.py
@@ -27,16 +27,12 @@ from pydantic_ai.models import Model
 
 from carapace.agent.loop import run_agent_turn as _run_agent_turn
 from carapace.git.store import GitStore
-from carapace.models import (
-    Config,
-    CredentialRegistryProtocol,
-    Deps,
-    SessionAttributes,
-    SessionState,
-    SkillInfo,
-    ToolCallCallback,
-    ToolResult,
-)
+from carapace.models.agent import Deps
+from carapace.models.config import Config
+from carapace.models.credentials import CredentialRegistryProtocol
+from carapace.models.session import SessionAttributes, SessionState
+from carapace.models.skills import SkillInfo
+from carapace.models.tooling import ToolCallCallback, ToolResult
 from carapace.notifications.router import NotificationRouter
 from carapace.sandbox.manager import SandboxManager
 from carapace.sandbox.runtime import SkillActivationInputs, SkillFileCredential
