+++ src/carapace/session/model_selection.py
@@ -11,7 +11,8 @@ from pydantic_ai.models import Model, infer_model
 from pydantic_ai.settings import ModelSettings
 
 from carapace.llm import model_settings_for_config
-from carapace.models import AvailableModelEntry, Config, SessionState, agent_available_model_entries
+from carapace.models.config import AvailableModelEntry, Config, agent_available_model_entries
+from carapace.models.session import SessionState
 from carapace.session.manager import SessionManager
 from carapace.session.types import ActiveSession
