+++ src/carapace/session/types.py
@@ -15,7 +15,8 @@ from typing import Any, Protocol, runtime_checkable
 from pydantic_ai.messages import ModelMessage
 from pydantic_ai.models import Model
 
-from carapace.models import SessionState, ToolResult
+from carapace.models.session import SessionState
+from carapace.models.tooling import ToolResult
 from carapace.security.context import ApprovalSource, ApprovalVerdict, SessionSecurity
 from carapace.security.sentinel import Sentinel
 from carapace.usage import LlmRequestLog, LlmRequestState, UsageTracker
